refactor: 알림 컴포넌트 세부화
