+++ index.html
@@ -17,7 +17,7 @@
     <link rel="icon" type="image/png" sizes="32x32" href="/favicon/favicon-32x32.png" />
     <link rel="apple-touch-icon" sizes="180x180" href="/favicon/apple-touch-icon.png" />
     -->
-    <link rel="icon" type="image/png" href="favicon/newcord_logo.jpg" />
+    <link rel="icon" type="image/png" href="favicon/nc_logo.png" />
     
     <!-- Google Font -->
     <link rel="preconnect" href="https://fonts.googleapis.com" />
