+++ package.json
@@ -46,6 +46,7 @@
     "react-dom": "^18.2.0",
     "react-helmet-async": "^1.3.0",
     "react-hook-form": "^7.47.0",
+    "react-intersection-observer": "^9.10.2",
     "react-kakao-maps-sdk": "^1.1.26",
     "react-router-dom": "^6.22.3",
     "react-simple-image-slider": "^2.4.1",
@@ -53,8 +54,8 @@
     "simplebar-react": "^3.2.4",
     "slick-carousel": "^1.8.1",
     "styled-components": "^6.1.8",
-    "zustand": "^4.5.0",
-    "vite-plugin-html": "^3.2.2"
+    "vite-plugin-html": "^3.2.2",
+    "zustand": "^4.5.0"
   },
   "devDependencies": {
     "@vitejs/plugin-react-swc": "^3.4.0",
