+++ src/layouts/dashboard/header.jsx
@@ -15,7 +15,7 @@ import Iconify from 'src/components/iconify';
 
 import { NAV, HEADER } from './config-layout';
 import AccountPopover from './common/account-popover';
-import NotificationsPopover from './common/notifications-popover';
+import NotificationsPopover from './common/NotificationsPopover';
 
 // ----------------------------------------------------------------------
